Make WebPortal a fast one!
We are working on this project and we want our visitors to have an awesome experience with our websites.
This portal we are working on is in word-press a great tool for web masters and SEOs to work upon.

But as it is known that it's hard to make things simple so as word-press.

I mean to say that word-press bit complex behind. Lots of CSS and js added with each plugin you add. And as we know that Google already instruct to use not more than 2 CSS files and all JS files should be combines to one. And for that you again have to install another plug-in.

But a word-press still favorite to web-master because despite its speed issue; you can do anything you want to do with it in few minutes.

You don't have to code and you don't have to invest your precious time in this.

So we are continue to making our projects in this but in near future we are thinking to change all these and will use our own framework for our own projects. Hope everything will go smooth and fine.

Thanks

rkgmxus@gmail.com